Refactored credential handling
