+++ system/types/credentials.go
@@ -5,7 +5,6 @@ import (
 
 	"github.com/cortezaproject/corteza-server/pkg/filter"
 	"github.com/jmoiron/sqlx/types"
-	"golang.org/x/crypto/bcrypt"
 )
 
 type (
@@ -33,23 +32,4 @@ type (
 
 func (u *Credentials) Valid() bool {
 	return u.ID > 0 && (u.ExpiresAt == nil || u.ExpiresAt.After(time.Now())) && u.DeletedAt == nil
-}
-
-// CompareHashAndPassword returns first valid credentials with matching hash
-func (cc CredentialsSet) CompareHashAndPassword(password string, validOnly bool) *Credentials {
-	for _, c := range cc {
-		if validOnly && !c.Valid() {
-			continue
-		}
-
-		if len(c.Credentials) == 0 {
-			continue
-		}
-
-		if bcrypt.CompareHashAndPassword([]byte(c.Credentials), []byte(password)) == nil {
-			return c
-		}
-	}
-
-	return nil
 }
